LPS-126038 Move DatasetContentRendererContextContributor to its own package
Since we have action, filter, and display, it makes sense to create a package
for content renderers, as they are the part of the renderer UI showing the
data items.
